Optimize Clairaudient and Tessellation with performance improvements, expand Transmutation to 32HP
Major performance optimizations:
- Clairaudient: Add parameter decimation (~20% CPU reduction), cached input checks (~8%), replace std::pow with exp2f (~10%), quantization result caching
- Tessellation: Add parameter decimation (~15-20% CPU reduction), external clock input with period measurement, smooth LED transitions using setBrightnessSmooth, fix double-speed LED blinking when clock connected

Module improvements:
- Transmutation: Expand from 30HP to 32HP, update knobs to ShapetakerKnobAlt variants for consistency, enlarge main screen to 100mm × 100mm
- Fix all build warnings: remove unused variables (getAttr, cols, rows), unnecessary lambda captures

Total expected CPU savings: 27-41% for Clairaudient, 15-20% for Tessellation with imperceptible latency

+++ src/ui/label_formatter.hpp
@@ -258,7 +258,7 @@ public:
         if (!module)
             return;
 
-        for (int i = 0; i < module->paramQuantities.size(); ++i) {
+        for (size_t i = 0; i < module->paramQuantities.size(); ++i) {
             rack::engine::ParamQuantity* pq = module->paramQuantities[i];
             if (!pq)
                 continue;
@@ -267,13 +267,13 @@ public:
 
         // Note: Input/Output port names are managed through configInput/configOutput
         // and stored in the module's inputInfos/outputInfos, not directly on the ports
-        for (int i = 0; i < (int)module->inputInfos.size(); ++i) {
+        for (size_t i = 0; i < module->inputInfos.size(); ++i) {
             if (module->inputInfos[i]) {
                 module->inputInfos[i]->name = normalizeInputLabel(module->inputInfos[i]->name);
             }
         }
 
-        for (int i = 0; i < (int)module->outputInfos.size(); ++i) {
+        for (size_t i = 0; i < module->outputInfos.size(); ++i) {
             if (module->outputInfos[i]) {
                 module->outputInfos[i]->name = normalizeOutputLabel(module->outputInfos[i]->name);
             }
